Extend Bylaws Revision Period
The initial bylaws for NPD allowed a simplified revision/amendment process for the first year of operations.  That year expires on June 19th and while we have made several revisions to the bylaws in the last few months, we still have at least two that need to happen as relates to off-year nominations and our own initial election process.

That's before we even touch on the fact that our Strategic Plan lists a thorough review of our bylaws as an organizational goal and most of our bylaws provisions have not been tested in production yet, or even critically examined since their initial drafting and review prior to our official formation.

We need more time to be able to tweak these without the full blown amendment process being necessary, and our membership is still small enough that this kind of notification and review process is superfluous.