Fix Amendment Loopholes
CAH fixed the SOP loophole in the LP Bylaws.  We exploited a similar loophole in the LPD Bylaws.  We fixed one such loophole in the NPD Bylaws related to national affiliations.

The better fix is to limit amendments themselves to any language requiring a higher vote threshold than just amending the threshold to require the higher threshold.