Set Default Meeting Time & Location
The bylaws require a quarterly meeting, but give the Board discretion to schedule it.  However, if the Board cannot reach a consensus on scheduling, how can the meeting take place?  The Bylaws establish a default date for the quarterly meetings on the third Sunday of the last month of the quarter, but do not specify a time or a location.  This amendment adds those details setting a time of noon and the location as the online forum established for other business.